update of types

+++ arduino/src/chimuino/chime.h
@@ -37,7 +37,7 @@ enum mode {
 
 struct Intention {
   enum mode what;
-  unsigned long when;  
+  uint32_t when;  
 };
 
 class IntentionProvider {
@@ -46,7 +46,7 @@ class IntentionProvider {
       // called when a GET <SOMETHING> was received
       // will receive as arguments the <SOMETHING> char*, 
       // and a pointer to serial port to use to answer
-      virtual Intention proposeNextMode(enum mode current_mode, unsigned long next_planned_action);
+      virtual Intention proposeNextMode(enum mode current_mode, uint32_t next_planned_action);
 
 };
 
@@ -76,10 +76,9 @@ class Chime: public BluetoothInformationProducer,
     void doDemoMedium();
     void doDemoStrong();
     
-    virtual Intention proposeNextMode(enum mode current_mode, unsigned long next_planned_action);
+    virtual Intention proposeNextMode(enum mode current_mode, uint32_t next_planned_action);
 
 
 };
 
 #endif // CHIME_H
-
